glwidget.py
# glwidget.py

from PyQt5.QtWidgets import QOpenGLWidget
from PyQt5.QtCore import Qt
from OpenGL.GL import *
from OpenGL.GLU import *
from shapes import Cube

class GLWidget(QOpenGLWidget):
    def __init__(self, parent=None):
        super(GLWidget, self).__init__(parent)
        self.rotation_angle = 0
        self.cube = Cube()  # Инициализация куба

    def initializeGL(self):
        glClearColor(0.0, 0.0, 0.0, 1.0)
        glEnable(GL_DEPTH_TEST)

        # Настройка перспективы
        glMatrixMode(GL_PROJECTION)
        gluPerspective(45.0, 1.0, 0.1, 100.0)
        glMatrixMode(GL_MODELVIEW)
        glLoadIdentity()
        gluLookAt(0, 0, 5, 0, 0, 0, 0, 1, 0)

    def resizeGL(self, w, h):
        glViewport(0, 0, w, h)
        glMatrixMode(GL_PROJECTION)
        glLoadIdentity()
        gluPerspective(45.0, w / h, 0.1, 100.0)
        glMatrixMode(GL_MODELVIEW)

    def paintGL(self):
        glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT)
        glLoadIdentity()
        glTranslatef(0.0, 0.0, -5.0)
        glRotatef(self.rotation_angle, 1.0, 1.0, 1.0)
        self.cube.draw()  # Рендеринг куба

    def rotate_object(self):
        self.rotation_angle += 1.0
        self.update()

    def keyPressEvent(self, event):
        if event.key() == Qt.Key_R:
            self.rotate_object()
