requirements.txt
PyOpenGL==3.1.7
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
